Addressing @nex3's concerns about the random() patch that @chriseppstein wrote. Also, wrote a large expansion to the test suite.
