+++ lib/sass/util.rb
@@ -1090,3 +1090,4 @@ end
 
 require 'sass/util/multibyte_string_scanner'
 require 'sass/util/normalized_map'
+require 'sass/util/random'
